+++ components/Tag.tsx
@@ -1,6 +1,5 @@
+import slugger from 'github-slugger';
 import Link from 'next/link';
-
-import kebabCase from '@/lib/kebabCase';
 
 interface Props {
   text: string;
@@ -8,7 +7,7 @@ interface Props {
 
 const Tag = ({ text }: Props) => {
   return (
-    <Link href={`/tags/${kebabCase(text)}`}>
+    <Link href={`/tags/${slugger.slug(text)}`}>
       <a className="mr-3 text-sm font-medium uppercase text-primary-500 hover:text-primary-600 dark:hover:text-primary-400">
         {text.split(' ').join('-')}
       </a>
