+++ data/blog/mdx-컴파일러-비교.mdx
@@ -1,7 +1,7 @@
 ---
 title: 'next-mdx-remote vs mdx-bundler'
 date: 2022-11-06T14:07:19.674Z
-lastmod: '2022-11-06'
+lastmod: 2022-11-07T14:07:19.674Z
 tags: ['블로그', 'mdx']
 draft: false
 category: 블로그 제작일지
