--- lib/kebabCase.ts
@@ -1,5 +1,0 @@
-import slugger from 'github-slugger';
-
-const kebabCase = (str: string) => slugger.slug(str);
-
-export default kebabCase;
