+++ pages/tags/index.tsx
@@ -1,8 +1,7 @@
-import { allBlogs } from 'contentlayer/generated';
+import slugger from 'github-slugger';
 import { GetStaticProps, InferGetStaticPropsType } from 'next';
 
-import { getAllTags } from '@/lib/contentlayer';
-import kebabCase from '@/lib/kebabCase';
+import { getAllTags } from '@/lib/getBlogInfo.mjs';
 
 import siteMetadata from '@/data/siteMetadata';
 
@@ -14,7 +13,7 @@ import Tag from '@/components/Tag';
 export const getStaticProps: GetStaticProps<{
   tags: Record<string, number>;
 }> = async () => {
-  const tags = await getAllTags(allBlogs);
+  const tags = (await getAllTags()) as Record<string, number>;
 
   return { props: { tags } };
 };
@@ -42,7 +41,7 @@ export default function Tags({
               <div key={t} className="mt-2 mb-2 mr-5">
                 <Tag text={t} />
                 <Link
-                  href={`/tags/${kebabCase(t)}`}
+                  href={`/tags/${slugger.slug(t)}`}
                   className="-ml-2 text-sm font-semibold uppercase text-gray-600 duration-500 dark:text-gray-300"
                 >
                   {` (${tags[t]})`}
